requirements.txt
gradio>=4.44.0
transformers>=4.44.0
accelerate>=0.33.0
torch>=2.2.0
faster-whisper>=1.0.0
pyttsx3>=2.90
elevenlabs>=1.50.4
soundfile>=0.12.1
pydub>=0.25.1
python-dotenv>=1.0.1
pydantic-settings>=2.4.0
spacy>=3.7.4
sentence-transformers>=3.0.1
loguru>=0.7.2
pytest>=8.3.2

